roles

+++ api.py
@@ -1,6 +1,6 @@
-from flask import Flask, render_template, request, redirect, flash, session
+from flask import Flask, render_template, request, redirect, flash, session, url_for, g
 from flaskext.mysql import MySQL
-from flask import jsonify # <- `jsonify` instead of `json`
+from flask import jsonify
 
 app = Flask(__name__)
 app.secret_key ="super secret key"
@@ -12,35 +12,98 @@ app.config['MYSQL_DATABASE_HOST'] = 'localhost'
 
 mysql = MySQL()
 mysql.init_app(app)
+ 
+@app.route('/', methods=['GET', 'POST'])
+def index():
+    if request.method == 'POST':
+        session.pop('user', None)
+        session.pop('admin', None)
 
+        if request.form['password'] == 'user':
+            session['user'] = request.form['username']
+            return redirect(url_for('protected'))
+
+        if request.form['password'] == 'admin':
+            session['admin'] = request.form['username']
+            return redirect(url_for('protected'))
+
+        if ((request.form['password'] == 'employee') and (request.form['username'] == 'employee')):
+            session['employee'] = request.form['username']
+            return redirect(url_for('protected'))
+
+
+    return render_template('index.html')
+
+@app.route('/protected')
+def protected():
+    if g.user or g.admin or g.employee:
+        return render_template('protected.html')
+
+    return redirect(url_for('index'))
+
+@app.before_request
+def before_request():
+    g.user = None
+    g.admin = None
+    g.employee = None 
+    if 'user' in session:
+        g.user = session['user']
+    if 'admin' in session:
+        g.admin = session['admin']
+    if 'employee' in session:
+        g.employee = session['employee']
+
+@app.route('/getsession')
+def getsession():
+    if 'user' in session:
+        return session['user']
+    if 'admin' in session:
+        return session['admin']
+    if 'employee' in session:
+        return session['employee']
+
+    return 'Not logged in!'
+
+@app.route('/dropsession')
+def dropsession():
+    session.pop('user', None)
+    session.pop('admin', None)
+    session.pop('employee', None)
+    return 'Dropped!'
 
   
 @app.route('/izdavanje')
 def izdavanje():
-    cursor = mysql.get_db().cursor()
-    cursor.execute('''SELECT * FROM izdavanje''')
-    data = cursor.fetchall()
-    return render_template('izdavanje.html', izdavanje=data)
-    #return str(data)
-    #return str(rv)
+    if g.user or g.admin or g.employee:
+        cursor = mysql.get_db().cursor()
+        cursor.execute('''SELECT * FROM izdavanje''')
+        data = cursor.fetchall()
+        return render_template('izdavanje.html', izdavanje=data)
+
+    return redirect(url_for('index'))
+
 
 @app.route('/izdavanje/<sifra>')
 def izdavanje_sifra(sifra):
-    cursor = mysql.connect().cursor()
-    cursor.execute('''select * FROM izdavanje where sifra = %s ''', (sifra))
-    r = [dict((cursor.description[i][0], value)
-         for i, value in enumerate(row)) for row in cursor.fetchall()]
-    return render_template('izdavanje_sifra.html', sifra=r[0])
+    if g.user or g.admin or g.employee:
+        cursor = mysql.connect().cursor()
+        cursor.execute('''select * FROM izdavanje where sifra = %s ''', (sifra))
+        r = [dict((cursor.description[i][0], value)
+             for i, value in enumerate(row)) for row in cursor.fetchall()]
+        return render_template('izdavanje_sifra.html', sifra=r[0])
+    return redirect(url_for('index'))
     #return str(r)
     #return jsonify({'results': r})
 
 @app.route('/izdavanje/obrisi/<int:sifra>')
 def izdavanje_obrisi(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("DELETE FROM izdavanje where sifra=%s", (sifra))
-    conn.commit()
-    return redirect('/izdavanje')
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("DELETE FROM izdavanje where sifra=%s", (sifra))
+        conn.commit()
+        return redirect('/izdavanje')
+    return redirect(url_for('index'))
 
 @app.route ('/izdavanje/dodaj')
 def izdavanje_dodaj():
@@ -48,33 +111,37 @@ def izdavanje_dodaj():
 
 @app.route('/dodaj_izdavanje', methods=['POST'])
 def insert_izdavanje():
-    if request.method == 'POST':
-        _datum_izdavanja = request.form['datum_izdavanja']
-        _datum_povratka = request.form['datum_povratka']
-        _cijena = request.form['cijena']
-        _izdavatelj = request.form['izdavatelj']
-        sql = '''INSERT INTO izdavanje(datum_izdavanja,datum_povratka,cijena,izdavatelj) VALUES (%s, %s, %s, %s)'''
-        data = (_datum_izdavanja, _datum_povratka, _cijena, _izdavatelj)
-        conn = mysql.connect()
-        cursor = conn.cursor()
-        cursor.execute(sql,data)
-        conn.commit()
-        #flash('user added successfully')
-        return redirect('/izdavanje')
-            #return render_template('/nakladnici')
-    else:
-            return 'Error adding user'
+    if g.admin or g.employee:
+        if request.method == 'POST':
+            _datum_izdavanja = request.form['datum_izdavanja']
+            _datum_povratka = request.form['datum_povratka']
+            _cijena = request.form['cijena']
+            _izdavatelj = request.form['izdavatelj']
+            sql = '''INSERT INTO izdavanje(datum_izdavanja,datum_povratka,cijena,izdavatelj) VALUES (%s, %s, %s, %s)'''
+            data = (_datum_izdavanja, _datum_povratka, _cijena, _izdavatelj)
+            conn = mysql.connect()
+            cursor = conn.cursor()
+            cursor.execute(sql,data)
+            conn.commit()
+            #flash('user added successfully')
+            return redirect('/izdavanje')
+                #return render_template('/nakladnici')
+        else:
+                return 'Error adding user'
+    return redirect (url_for('izdavanje'))
 
 @app.route('/izdavanje_edit/<int:sifra>')
 def edit_izdavanje_view(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("SELECT * FROM izdavanje where sifra=%s",sifra)
-    row = cursor.fetchone()
-    if row:
-        return render_template('izdavanje_edit.html', row=row)
-    else:
-        return 'Error loading #{id}'.format(sifra=sifra)
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("SELECT * FROM izdavanje where sifra=%s",sifra)
+        row = cursor.fetchone()
+        if row:
+            return render_template('izdavanje_edit.html', row=row)
+        else:
+            return 'Error loading #{id}'.format(sifra=sifra)
+    return redirect(url_for('izdavanje'))
 
 @app.route('/edit_izdavanje', methods=['POST'])
 def edit_izdavanje():
@@ -100,8 +167,7 @@ def izdavatelj():
     cursor.execute('''SELECT * FROM izdavatelj''')
     data = cursor.fetchall()
     return render_template('izdavatelj.html', izdavatelj=data)
-    #return str(data)
-    #return str(rv)
+
 
 @app.route('/izdavatelj/<sifra>')
 def izdavatelj_sifra(sifra):
@@ -115,16 +181,20 @@ def izdavatelj_sifra(sifra):
 
 @app.route('/izdavatelj/obrisi/<int:sifra>')
 def izdavatelj_obrisi(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("DELETE FROM izdavatelj where sifra=%s", (sifra))
-    conn.commit()
-    return redirect('/izdavatelj')
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("DELETE FROM izdavatelj where sifra=%s", (sifra))
+        conn.commit()
+        return redirect('/izdavatelj')
+    return redirect(url_for('izdavatelj'))
 
 
 @app.route ('/izdavatelj/dodaj')
 def izdavatelj_dodaj():
-    return render_template('izdavatelj_dodaj.html')
+    if g.admin or g.employee:
+        return render_template('izdavatelj_dodaj.html')
+    return redirect(url_for('izdavatelj'))
 
 @app.route('/dodaj_izdavatelj', methods=['POST'])
 def insert_izdavatelj():
@@ -148,14 +218,16 @@ def insert_izdavatelj():
 
 @app.route('/izdavatelj_edit/<int:sifra>')
 def izdavatelj_edit_view(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("SELECT * FROM izdavatelj where sifra=%s",sifra)
-    row = cursor.fetchone()
-    if row:
-        return render_template('izdavatelj_edit.html', row=row)
-    else:
-        return 'Error loading #{id}'.format(sifra=sifra)
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("SELECT * FROM izdavatelj where sifra=%s",sifra)
+        row = cursor.fetchone()
+        if row:
+            return render_template('izdavatelj_edit.html', row=row)
+        else:
+            return 'Error loading #{id}'.format(sifra=sifra)
+    return redirect(url_for('izdavatelj'))
 
 @app.route('/edit_izdavatelj', methods=['POST'])
 def edit_izdavatelj():
@@ -177,7 +249,6 @@ def edit_izdavatelj():
         return 'Error while updating user'
 
 
-
 @app.route('/knjiga')
 def knjiga():
     cursor = mysql.get_db().cursor()
@@ -187,7 +258,9 @@ def knjiga():
 
 @app.route ('/knjiga/dodaj')
 def dodaj_knjiga_view():
-    return render_template('knjiga_dodaj.html')
+    if g.admin or g.employee:
+        return render_template('knjiga_dodaj.html')
+    return redirect(url_for('knjiga'))
 
 @app.route('/dodaj_knjiga', methods=['POST'])
 def insert_knjiga():
@@ -212,14 +285,16 @@ def insert_knjiga():
 
 @app.route('/knjiga_edit/<int:sifra>')
 def edit_view(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("SELECT * FROM knjiga where sifra=%s",sifra)
-    row = cursor.fetchone()
-    if row:
-        return render_template('knjiga_edit.html', row=row)
-    else:
-        return 'Error loading #{id}'.format(sifra=sifra)
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("SELECT * FROM knjiga where sifra=%s",sifra)
+        row = cursor.fetchone()
+        if row:
+            return render_template('knjiga_edit.html', row=row)
+        else:
+            return 'Error loading #{id}'.format(sifra=sifra)
+    return redirect(url_for('knjiga'))
 
 @app.route('/update', methods=['POST'])
 def edit_knjiga():
@@ -243,11 +318,13 @@ def edit_knjiga():
 
 @app.route('/knjiga/obrisi/<int:sifra>')
 def obrisi_knjiga(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("DELETE FROM knjiga where sifra=%s", (sifra))
-    conn.commit()
-    return redirect('/knjiga')
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("DELETE FROM knjiga where sifra=%s", (sifra))
+        conn.commit()
+        return redirect('/knjiga')
+    return redirect(url_for('knjiga'))
 
 @app.route('/knjiga/<sifra>')
 def knjiga_sifra(sifra):
@@ -277,43 +354,51 @@ def nakladnici_sifra(nakladnici_sifra):
 
 @app.route ('/nakladnici/dodaj')
 def dodaj_nakladnik_view():
+    if g.admin or g.employee:
+        return redirect(url_for('knjiga'))
     return render_template('nakladnici_dodaj.html')
 
 @app.route('/nakladnici_dodaj', methods=['POST'])
 def insert_nakladnici():
-    if request.method == 'POST':
-        _naziv = request.form['Naziv']
-        _mjesto = request.form['Mjesto']
-        sql = '''INSERT INTO nakladnik(Naziv,Mjesto) VALUES (%s, %s)'''
-        data = (_naziv, _mjesto)
-        conn = mysql.connect()
-        cursor = conn.cursor()
-        cursor.execute(sql,data)
-        conn.commit()
-        #flash('user added successfully')
-        return redirect('/nakladnici')
-            #return render_template('/nakladnici')
-    else:
-            return 'Error adding user'
+    if g.admin or g.employee:
+        if request.method == 'POST':
+            _naziv = request.form['Naziv']
+            _mjesto = request.form['Mjesto']
+            sql = '''INSERT INTO nakladnik(Naziv,Mjesto) VALUES (%s, %s)'''
+            data = (_naziv, _mjesto)
+            conn = mysql.connect()
+            cursor = conn.cursor()
+            cursor.execute(sql,data)
+            conn.commit()
+            #flash('user added successfully')
+            return redirect('/nakladnici')
+                #return render_template('/nakladnici')
+        else:
+                return 'Error adding user'
+    return redirect(url_for('nakladnici'))
 
 @app.route('/nakladnici/obrisi/<int:sifra>')
 def nakladnici_obrisi(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("DELETE FROM nakladnik where sifra=%s", (sifra))
-    conn.commit()
-    return redirect('/nakladnici')
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("DELETE FROM nakladnik where sifra=%s", (sifra))
+        conn.commit()
+        return redirect('/nakladnici')
+    return redirect(url_for('nakladnici'))
 
 @app.route('/nakladnici_edit/<int:sifra>')
 def edit_nakladnici_view(sifra):
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    cursor.execute("SELECT * FROM nakladnik where sifra=%s",sifra)
-    row = cursor.fetchone()
-    if row:
-        return render_template('nakladnici_edit.html', row=row)
-    else:
-        return 'Error loading #{id}'.format(sifra=sifra)
+    if g.admin or g.employee:
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        cursor.execute("SELECT * FROM nakladnik where sifra=%s",sifra)
+        row = cursor.fetchone()
+        if row:
+            return render_template('nakladnici_edit.html', row=row)
+        else:
+            return 'Error loading #{id}'.format(sifra=sifra)
+    return redirect(url_for('nakladnici'))
 
 @app.route('/update_nakladnik', methods=['POST'])
 def edit_nakladnici():
